% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DetectChangePoint_NumUnknown.R
\name{detect_estimated_cp}
\alias{detect_estimated_cp}
\title{Title}
\usage{
detect_estimated_cp(
  X = NULL,
  D = NULL,
  d = NULL,
  dist.method = "average",
  lambda = 0.02
)
}
\arguments{
\item{lambda}{}
}
\value{

}
\description{
Title
}
