% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distances.R
\name{distmat_HDLSS}
\alias{distmat_HDLSS}
\title{Distance Matrix Calculation using genMADD or MADD distance}
\usage{
distmat_HDLSS(X, option = c("genMADD", "MADD"))
}
\arguments{
\item{X}{Data matrix (n X p) where n denotes number of observations. Each row is a p dimensional observation vector.}

\item{option}{Distance function to use. Should be one of "genMADD" or "MADD". The default is "genMADD".}
}
\value{
Returns a (n X n) matrix containing the distance between the observations.
}
\description{
Distance Matrix Calculation using genMADD or MADD distance
}
\examples{
set.seed(1)
# Generating n = 10 observations from t distribution with dimension p = 5
X = matrix(stats::rt((10*5), ncp = 0, df = 4), nrow = 10, ncol = 5)

# Distance matrix
D_mat1 = distmat_HDLSS(X = X) # using default generalized MADD distance ("genMADD")
D_mat2 = distmat_HDLSS(X = X, option = "MADD") # using MADD distance ("MADD")
}
