% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultipleChangePoint_NumKnown.R
\name{detect_multiple_cp}
\alias{detect_multiple_cp}
\title{Title}
\usage{
detect_multiple_cp(X = NULL, D = NULL, numcp, dist.method = "average")
}
\arguments{
\item{dist.method}{}
}
\value{

}
\description{
Title
}
